Refactored the routes using Express router
